TASK: Add DLX arguments to declared queue and bind to DLX
If the configuration argument `useDLX` is of type string it's now
assumed as the name of the DLX.
The queue argument `x-dead-letter-exchange` is set to the value of
`useDLX` and the queue argument `x-dead-letter-routing-key` is set
to the queue name prefixed with `queue.`.

If `declare` is truthy, the queue is bound to the `amq.direct`.